Separate code into separate files
