+++ src/native_driver.jl
@@ -1,132 +1,3 @@
-const EMPTY_POD =
-    json(Dict("kind" => "Pod",
-              "metadata" => Dict(),
-              "spec" => Dict("restartPolicy" => "Never",
-                             "tolerations" => [],
-                             "containers" => [],
-                             "affinity" => Dict())))
-
-const DEFAULT_NAMESPACE = "default"
-const NAMESPACE_FILE = "/var/run/secrets/kubernetes.io/serviceaccount/namespace"
-
-
-"""
-    config_namespace() -> Union{String,Nothing}
-
-Determine the Kubernetes namespace as specified by the current config context. If the
-namespace is not set, the current context is not set, or the current context is not defined
-then `nothing` will be returned.
-"""
-function config_namespace()
-    # https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces/
-    #
-    # Equivalent to running `kubectl config view --minify --output='jsonpath={..namespace}'`
-    # but improves handling of corner cases.
-    kubectl() do exe
-        context = read(`$exe config view --output='jsonpath={.current-context}'`, String)
-        isempty(context) && return nothing
-
-        # Note: The output from `kubectl config view` reports a missing `namespace` entry,
-        # `namespace: null`, and `namespace: ""` as the same.
-        output = "jsonpath={.contexts[?(@.name=='$context')].context.namespace}"
-        namespace = read(`$exe config view --output=$output`, String)
-        return !isempty(namespace) ? namespace : nothing
-    end
-end
-
-
-"""
-    pod_namespace() -> Union{String,Nothing}
-
-Determine the namespace of the pod if running inside of a Kubernetes pod, otherwise return
-`nothing`.
-"""
-function pod_namespace()
-    return if @mock isfile(NAMESPACE_FILE)
-        @mock read(NAMESPACE_FILE, String)
-    else
-        nothing
-    end
-end
-
-
-"""
-    current_namespace() -> String
-
-Determine the Kubernetes namespace as specified by the current config or, when running
-inside a pod, the namespace of the pod. If the namespace is cannot be determined the default
-namespace ("$DEFAULT_NAMESPACE") will be returned.
-"""
-function current_namespace()
-    namespace = config_namespace()
-    namespace !== nothing && return namespace
-
-    namespace = pod_namespace()
-    namespace !== nothing && return namespace
-
-    return DEFAULT_NAMESPACE
-end
-
-
-"""
-    self_pod(ctx::KuberContext)
-
-Kuber object representing the pod this julia session is running in.
-"""
-function self_pod(ctx)
-    # The following code is equivalent to calling Kuber's `get(ctx, :Pod, ENV["HOSTNAME"])`
-    # but reduces noise by avoiding nested rethrow calls.
-    # Fixed in Kuber.jl in: https://github.com/JuliaComputing/Kuber.jl/pull/26
-    isempty(ctx.apis) && Kuber.set_api_versions!(ctx)
-    api_ctx = Kuber._get_apictx(ctx, :Pod, nothing)
-    return Kuber.readNamespacedPod(api_ctx, ENV["HOSTNAME"], ctx.namespace)
-end
-
-
-"""
-    default_pod(ctx, port, cmd::Cmd, driver_name::String; image=nothing, memory::String="4Gi", cpu::String="1", base_obj=kuber_obj(ctx, EMPTY_POD))
-
-Kuber object representing a pod with a single worker container.
-
-If `isnothing(image) == true`, the driver pod is required to have a single container, whose image will be used.
-"""
-function default_pod(ctx, port, cmd::Cmd, driver_name::String; image=nothing, memory::String="4Gi", cpu::String="1", serviceAccountName=nothing, base_obj=kuber_obj(ctx, EMPTY_POD), kwargs...)
-    ko = base_obj
-    ko.metadata.name = "$(driver_name)-worker-$port"
-    cmdo = `$cmd --bind-to=0:$port`
-    if isnothing(image)
-        self = self_pod(ctx)
-        if length(self.spec.containers) > 1
-            error("`default_pod` called with `image = nothing`, driver pod must contain a single container: use `configure` closure kwarg to set worker images manually.")
-        end
-        image = last(self.spec.containers).image
-    end
-    push!(ko.spec.containers,
-          json(Dict("name" => "$(driver_name)-worker-$port",
-                    "image" => image,
-                    "command" => collect(cmdo),
-                    "resources" => Dict("requests" => Dict("memory" => memory,
-                                                           "cpu" => cpu)),
-                                        "limit"    => Dict("memory" => memory,
-                                                           "cpu" => cpu))))
-    if !isnothing(serviceAccountName)
-        ko.spec.serviceAccountName = serviceAccountName
-    end
-    return ko
-end
-
-function default_pods_and_context(namespace=current_namespace(); configure, ports, driver_name::String="driver", cmd::Cmd=`julia $(worker_arg())`, kwargs...)
-    ctx = KuberContext()
-    Kuber.set_api_versions!(ctx; verbose=false)
-    set_ns(ctx, namespace)
-
-    # Avoid using a generator with `Dict` as any raised exception would be displayed twice:
-    # https://github.com/JuliaLang/julia/issues/33147
-    pods = Dict([port => configure(default_pod(ctx, port, cmd, driver_name; kwargs...)) for port in ports])
-
-    return pods, ctx
-end
-
 struct K8sNativeManager <: ClusterManager
     ctx::Any
     pods::Dict{Int, Any}
